Change`Pid` and `sid` to save 2 bytes per use
Saves 8*1146 bytes of memory in `TextStyleProperty` alone

+++ src/engraving/compat/dummyelement.h
@@ -25,7 +25,7 @@
 #include "../dom/engravingitem.h"
 
 namespace mu::engraving {
-enum class Pid : int;
+enum class Pid : short;
 
 class RootItem;
 }
